+++ server/accounts.json
@@ -1,1 +1,1 @@
-{"accounts":[{"login":"zhenia","password":"$2a$12$1IAysIfsXqUA80jPnVszAO.Pudhn5g8rlLXAFzGeeShupcizKG6MO","profile":{"avatar":"","username":"Geshka","description":"Hello world!"}}]}
+{"accounts":[{"login":"zhenia","password":"$2a$12$ofGIrVe3JyUDNGARz7ANv.Gf9vSbncCbkx/PX/YM/vhLVFdKxFg6e","profile":{"avatar":"","username":"Geshka","description":"Hello world!"}}]}
