+++ src/api/components/pages/homePage/categories.scss
