+++ src/api/components/pages/homePage/home.tsx
@@ -1,16 +1,16 @@
 import { useState, useEffect, ChangeEvent, SetStateAction } from "react";
 // import debounce from "debounce";
-import debounce from "@/api/debounse";
+import debounce from "@/helpers/debounse";
 import getData from "@/api/apiSearchData";
 import getRecentlyAddedProducts from "@/api/getTopProducts";
-import loadingImage from "@/assets/images/loading.svg";
+import loadingImage from "@/api/assets/images/loading.svg";
 
 // categories
-import CardsContainer from "@/components/cards/cardsContainer";
+import CardsContainer from "@/api/components/cards/cardsContainer";
 // import Card from "@/Components/Cards/card";
 
-import "@/pages/homePage/home.scss";
-import "@/pages/homePage/categories.scss";
+import "@/api/components/pages/homePage/home.scss";
+import "@/api/components/pages/homePage/categories.scss";
 
 import Categories from "./categories";
