+++ src/styles/pages.scss
@@ -1,7 +1,7 @@
 .pages_container {
   padding: 10px;
   min-height: 80vh;
-  background: $background url("@/assets/images/background1.jpg") no-repeat center fixed;
+  background: $background url("@/api/assets/images/background1.jpg") no-repeat center fixed;
 }
 
 .page_content_container {
